Readkeys now works and sketch updated for new printing functions.
The main code now prints any output to the LCD and incoming serial port,
which then shows up over the serial connection. Readkeys uses a thread
to read the incoming text, so it does not block the text input from the
keyboard. There does still seem to be a serial port race condition that
happens, but it only seems to occasionally show up.