Put some of the nrf code into a function and made serial echo optional
The nrf24 transmit and receive code is now in a function to reduce the
code size. Added the neenable variable which can be used to toggle
whether the keycodes are echoed back over the serial connection from the
Arduino. Also added a bit of mutex code to the readkeys code, but it
probably does not make much of a difference